Moved Eve data-related files into subfolder in $lib/
