+++ src/lib/eve-data/EveCharacter.ts
