+++ src/lib/eve-data/EveDogma.ts
