+++ src/lib/eve-data/EveIndustry.ts
